releasing: Add example commands to release directions
Note that for weston, configure.ac needs version numbers updated in
several places.  `make distcheck` for weston no longer triggers the test
suite to run, so update the directions to run a `make check`
pre-release.  Switch to using the X.org release.sh script in place of
some of the manual steps.

Signed-off-by: Bryce Harrington <bryce@osg.samsung.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>